Commit message: Switched CI triggers from automatic to manual
The GitHub action trigger for Maven CI workflow has been changed from automatic to manual. The file 'maven.yml' was deleted and replaced with 'ci-manually.yml'. This update is implemented to provide more control over when the workflow is executed, reducing unnecessary usage of resources when small or insignificant changes are pushed to the main branch.